Seperate keyInputHandler class
